co2-sensor/requirements.txt
autopep8==1.6.0
pyserial==3.5
python-dotenv==0.20.0
